requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langchain-tavily>=0.0.1
python-dotenv==1.0.1